requirements.txt
torch>=1.10.0
torchvision>=0.11.0
timm>=0.6.0
numpy>=1.20.0
Pillow>=8.0.0
opencv-python>=4.5.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.60.0
requests>=2.25.0
gradio>=3.0.0
ultralytics>=8.0.0
